requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
a2wsgi>=1.10.10
SQLAlchemy>=2.0.40
psycopg2-binary>=2.9.10
pydantic-settings>=2.10.1
python-jose[cryptography]>=3.5.0
passlib>=1.7.4
email-validator>=2.2.0
jinja2>=3.1.6
python-multipart>=0.0.20
cloudinary>=1.44.1
reportlab>=4.4.0
